requirements.txt
# MillionTwigs — Satellite Vegetation Analysis
# Python 3.10+

# ── Core geospatial ────────────────────────────────────────────────────────────
rasterio>=1.3.9
gdal>=3.6.0          # system GDAL must be installed first: apt install libgdal-dev
geopandas>=0.14.3
shapely>=2.0.3
pyproj>=3.6.1
fiona>=1.9.5
rtree>=1.2.0

# ── Satellite data access ──────────────────────────────────────────────────────
earthengine-api>=0.1.390     # Google Earth Engine Python API
sentinelsat>=1.2.1           # Copernicus Open Access Hub / Dataspace
landsatxplore>=0.13          # USGS EarthExplorer Landsat browser

# ── Image processing ───────────────────────────────────────────────────────────
numpy>=1.26.4
scipy>=1.12.0
scikit-image>=0.22.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0

# ── Machine learning ───────────────────────────────────────────────────────────
scikit-learn>=1.4.2
xgboost>=2.0.3

# ── Deep learning (tree detection & segmentation) ─────────────────────────────
torch>=2.2.2
torchvision>=0.17.2
segmentation-models-pytorch>=0.3.3   # U-Net, FPN, DeepLab with pretrained encoders
albumentations>=1.4.3                # augmentation for segmentation training

# ── DeepForest — individual tree crown detection (Weinstein et al. 2020) ──────
deepforest>=1.3.3

# ── Segment Anything Model (SAM) — Meta 2023 ──────────────────────────────────
# Install separately: pip install git+https://github.com/facebookresearch/segment-anything.git
# segment-anything

# ── Vegetation indices & remote sensing utilities ─────────────────────────────
# py6s>=1.9.2        # atmospheric correction (requires 6S binary)
pysptools>=0.15.0   # hyperspectral & spectral unmixing utilities

# ── Data science ───────────────────────────────────────────────────────────────
pandas>=2.2.1
matplotlib>=3.8.4
seaborn>=0.13.2
plotly>=5.20.0

# ── Visualisation & reporting ─────────────────────────────────────────────────
folium>=0.16.0         # interactive Leaflet maps in Python
branca>=0.7.2
contextily>=1.5.1      # basemaps for matplotlib
fpdf2>=2.7.9           # PDF report generation

# ── Notebook support ──────────────────────────────────────────────────────────
jupyter>=1.0.0
jupyterlab>=4.1.5
ipywidgets>=8.1.2
ipyleaflet>=0.18.2

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.2
pyyaml>=6.0.1
click>=8.1.7
requests>=2.31.0
python-dotenv>=1.0.1
